[updated] udpated PrefixHeader

+++ AZS/AZS/SupportFile/AZSPrefixHeader.swift
@@ -22,10 +22,10 @@ let KScreenHeight = UIScreen.main.bounds.height
 
 let KDevice_iPhoneXSeries = (__CGSizeEqualToSize(CGSize.init(width: 375.0, height: 812.0), UIScreen.main.bounds.size) || __CGSizeEqualToSize(CGSize.init(width: 414.0, height: 896.0), UIScreen.main.bounds.size))
 
-let KNavHeight = (KDevice_iPhoneXSeries ? 88.0 : 64.0)
-let KTabBarHeight = (KDevice_iPhoneXSeries ? 83.0 : 49.0)
-let KStatusBarHeight = (KDevice_iPhoneXSeries ? 44.0 : 20.0)
-let KBottomMargin = (KDevice_iPhoneXSeries ? 34.0 : 0)
+let KNavHeight = CGFloat(KDevice_iPhoneXSeries ? 88.0 : 64.0)
+let KTabBarHeight = CGFloat(KDevice_iPhoneXSeries ? 83.0 : 49.0)
+let KStatusBarHeight = CGFloat(KDevice_iPhoneXSeries ? 44.0 : 20.0)
+let KBottomMargin = CGFloat(KDevice_iPhoneXSeries ? 34.0 : 0)
 
 let TT_IID = "17769976909"
 let TT_DEVICE_ID = "41312231473"
